supplychain-backend/indexer/requirements.txt
web3==6.4.0
python-dotenv==1.0.0
